requirements.txt
beautifulsoup4
bs4
certifi
cffi
charset-normalizer
click
colorama
cryptography
Deprecated
Flask
gcloud
googleapis-common-protos
gunicorn
httplib2
idna
itsdangerous
Jinja2
jwcrypto
lxml
MarkupSafe
oauth2client
protobuf
pyasn1
pyasn1-modules
pycparser
pycryptodome
pyparsing
Pyrebase4
python-jwt
rsa
six
soupsieve
urllib3
Werkzeug
wrapt
